Added arity options for functions
Now the "old" implementation has the ability to do unary operations.
Functions can be given an arity in the arity-table hash, and user-made
functions can be given implementations in the func-table hash. Arity
is given in the same place regardless.

stacket> o
stacket> (4 ~ 5 +)
'(1)

(Note: ~ is currently the unary negation function.)