Fixed execution
There was an error this whole time with the function application. I was
overwriting the entire stack for each function call. I didn't see this
at first because for simple maths it's fine to only accumulate one
value. However, now the following is an option:

stacket> o
stacket> ("mass?" prompt "acceleration?" prompt * "force is")
mass?
5.5
acceleration?
4.5
'("force is" 24.75)